pyproject.toml
[project]
name = "hstack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.135.1",
    "google-genai>=1.67.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
